+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move utils folder content in assets folder (#210)
 - Increase tocdepth in SDK API reference to display methods under Client (#219)
 - Update substratools Docker image name (#222)
+- Use the new `add_metric` SubstraFL function in examples (#214)
 
 ## [0.22.0]